updated database script

+++ download_DBs.sh
@@ -5,44 +5,55 @@ mkdir -p databases
 cd databases
 
 # Pfam database
-wget ftp://ftp.ebi.ac.uk/pub/databases/Pfam/releases/Pfam31.0/Pfam-A.hmm.gz
-gunzip Pfam-A.hmm.gz
-
+if [[ ! -s Pfam-A.hmm ]]; then
+  wget ftp://ftp.ebi.ac.uk/pub/databases/Pfam/releases/Pfam31.0/Pfam-A.hmm.gz
+  gunzip Pfam-A.hmm.gz
+fi
 # Pfam clans
-wget ftp://ftp.ebi.ac.uk/pub/databases/Pfam/current_release/Pfam-A.clans.tsv.gz
-
+if [[ ! -s Pfam-A.clans.tsv.gz ]]; then
+  wget ftp://ftp.ebi.ac.uk/pub/databases/Pfam/current_release/Pfam-A.clans.tsv.gz
+fi
 # Pfam HHBLITS DB
-wget http://wwwuser.gwdg.de/~compbiol/data/hhsuite/databases/hhsuite_dbs/pfamA_31.0.tgz
-tar xvfz pfamA_31.0.tgz
+if [[ ! -s pfam_hhm.ffdata ]]; then
+  wget http://wwwuser.gwdg.de/~compbiol/data/hhsuite/databases/hhsuite_dbs/pfamA_31.0.tgz
+  tar xvfz pfamA_31.0.tgz
+fi
 
 # Pfam list common domain terms
-wget https://ndownloader.figshare.com/files/23756204 -O Pfam-31_names_mod_01122019.tsv
-
+if [[ ! -s Pfam-31_names_mod_01122019.tsv ]]; then
+  wget https://ndownloader.figshare.com/files/23756204 -O Pfam-31_names_mod_01122019.tsv
+fi
 # Antifam databases
-wget ftp://ftp.ebi.ac.uk/pub/databases/Pfam/AntiFam/current/Antifam.tar.gz
-tar xvfz Antifam.tar.gz
-../bin/hmmpress AntiFam.hmm
-
+if [[ ! -s AntiFam.hmm ]]; then
+  wget ftp://ftp.ebi.ac.uk/pub/databases/Pfam/AntiFam/current/Antifam.tar.gz
+  tar xvfz Antifam.tar.gz
+  ../bin/hmmpress AntiFam.hmm
+fi
 # Uniref90
-wget ftp://ftp.uniprot.org/pub/databases/uniprot/uniref/uniref90/uniref90.fasta.gz
-# Create the protein description file:
-zcat uniref90.fasta.gz | grep '^>' | sed 's/^>//' | sed 's/ /\t/' | sed 's/ /_/g' | gzip > uniref90.proteins.tsv.gz
-
+if [[ ! -s uniref90.fasta.gz ]]; then
+  aria2c --file-allocation=none -c -x 10 -s 10 ftp://ftp.uniprot.org/pub/databases/uniprot/uniref/uniref90/uniref90.fasta.gz
+  # Create the protein description file:
+  zcat uniref90.fasta.gz | grep '^>' | sed 's/^>//' | sed 's/ /\t/' | sed 's/ /_/g' | gzip > uniref90.proteins.tsv.gz
+fi
 
 # NCBI nr
-wget ftp://ftp.ncbi.nih.gov/blast/db/FASTA/nr.gz
-mv nr.gz nr.fasta.gz
-# Create the protein description file:
-zcat nr.fasta.gz | grep '^>' | sed 's/^>//' | sed 's/ /\t/' | sed 's/ /_/g' | gzip > nr.proteins.tsv.gz
-
+if [[ ! -s nr.fasta.gz ]]; then
+  aria2c --file-allocation=none -c -x 10 -s 10 ftp://ftp.ncbi.nih.gov/blast/db/FASTA/nr.gz
+  mv nr.gz nr.fasta.gz
+  # Create the protein description file:
+  zcat nr.fasta.gz | grep '^>' | sed 's/^>//' | sed 's/ /\t/' | sed 's/ /_/g' | gzip > nr.proteins.tsv.gz
+fi
 # Uniclust HHBLITS DB
-wget http://wwwuser.gwdg.de/~compbiol/uniclust/2018_08/uniclust30_2018_08.tar.gz
-tar xvfz uniclust30_2018_08.tar.gz
+if [[ ! -s uniclust30_2018_08/uniclust30_2018_08_a3m.ffdata ]]; then
+  aria2c --file-allocation=none -c -x 10 -s 10 http://wwwuser.gwdg.de/~compbiol/uniclust/2018_08/uniclust30_2018_08.tar.gz
+  tar xvfz uniclust30_2018_08.tar.gz
+fi
 
 # DPD and info
-wget https://ndownloader.figshare.com/files/23756312 -O dpd_uniprot_sprot.fasta.gz
-wget https://ndownloader.figshare.com/files/23756306 -O dpd_ids_all_info.tsv.gz
-
+if [[ ! -s dpd_uniprot_sprot.fasta.gz ]]; then
+  wget https://ndownloader.figshare.com/files/23756312 -O dpd_uniprot_sprot.fasta.gz
+  wget https://ndownloader.figshare.com/files/23756306 -O dpd_ids_all_info.tsv.gz
+fi
 # Uniprot KB
 # wget ftp://ftp.expasy.org/databases/uniprot/current_release/knowledgebase/complete/uniprot_trembl.fasta.gz
 # wget wget ftp://ftp.expasy.org/databases/uniprot/current_release/knowledgebase/complete/uniprot_sprot.fasta.gz
